+++ bot/bot_factory.py
@@ -65,8 +65,8 @@ def create_bot(bot_type):
         return ZHIPUAIBot()
 
     elif bot_type == const.COZE:
-        from bot.coze.coze_bot import CozeBot
-        return CozeBot()
+        from bot.bytedance.bytedance_coze_bot import ByteDanceCozeBot
+        return ByteDanceCozeBot()
 
     elif bot_type == const.MOONSHOT:
         from bot.moonshot.moonshot_bot import MoonshotBot
